requirements.txt
adapter-transformers==3.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
filelock==3.9.0
huggingface-hub==0.12.1
idna==3.4
joblib==1.2.0
nltk==3.8.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.1
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.13.1
tqdm==4.64.1
typing-extensions==4.5.0
urllib3==1.26.14
